It compiles!!!!
Huge refactor. Moved the location of the dynamic route such that the view-deck and suggestion page are both nested underneath it. Was getting errors since neither of the deck context providers were the default exports and therefore could not have children. Renamed a few exports, split the logic of contextprovider into helper file
