+++ src/app/(deck-changers)/[slug]/layout.jsx
@@ -1,0 +1,8 @@
+import UserDeckContextProvider from "@/components/context-providers/userDeckContextProvider"
+
+export default function DeckChangersLayout({ children }){
+
+    return(
+        <UserDeckContextProvider>{children}</UserDeckContextProvider>
+    )
+}
