+++ src/app/(deck-changers)/[slug]/view-import-deck/layout.jsx
@@ -1,0 +1,10 @@
+import  ViewDeckContextProvider  from '@/components/context-providers/viewDeckContextProvider';
+
+export default async function ViewDeckLayout({children}){
+
+    return(
+        <ViewDeckContextProvider>
+            {children}
+        </ViewDeckContextProvider>
+    )
+}
